refactor(mentor): 일부 버튼들 공통으로 수정
