+++ src/pages/Homepage.css
@@ -19,6 +19,10 @@
   align-items: center;
 }
 
+.hero-text a {
+  text-decoration: none;
+}
+
 .hero-title {
   width: 100%;
   font-weight: 700;
